Revert "api code implementation"
